+++ src/errors.js
@@ -33,13 +33,34 @@ export class ValueNotFoundError extends Error {
   }
 }
 
+/*::
+export type ExpectedValueOrder =
+  | 'less than'
+  | 'greater than'
+  | 'equal to';
+*/
+
 export class ValueOrderError extends Error {
-  constructor(expected/*: mixed */, got/*: mixed */) {
+  /*::
+  +v1: mixed;
+  +v2: mixed;
+  +expectedOrder: ExpectedValueOrder;
+  */
+
+  constructor(
+    v1/*: mixed */,
+    v2/*: mixed */,
+    expectedOrder/*: ExpectedValueOrder */,
+  ) {
     super(
-      'The expected order of the value has changed: ' +
-      String(expected) + ' (expected) ' +
-      'vs. ' +
-      String(got) + ' (got)',
+      'The relative order of values has changed: expected ' +
+      String(v1) + ' to be ' +
+      expectedOrder + ' ' +
+      String(v2),
     );
+    this.name = 'ValueOrderError';
+    this.v1 = v1;
+    this.v2 = v2;
+    this.expectedOrder = expectedOrder;
   }
 }